chore: Rename the mixin to json-api-bagaaravel-serializer
