+++ tests/unit/mixins/relationship-support-serializer-test.js
@@ -1,4 +1,4 @@
-import JSONAPISerializerMixin from '@bagaar/ember-data-bagaaravel/mixins/json-api-serializer';
+import JSONAPIBagaaravelSerializerMixin from '@bagaar/ember-data-bagaaravel/mixins/json-api-bagaaravel-serializer';
 import RelationshipSupportSerializerMixin from '@bagaar/ember-data-bagaaravel/mixins/relationship-support-serializer';
 import JSONAPIAdapter from 'ember-data/adapters/json-api';
 import JSONAPISerializer from 'ember-data/serializers/json-api';
@@ -19,7 +19,7 @@ module('Unit | Mixin | relationship-support-serializer', function (hooks) {
     });
 
     let UserSerializer = JSONAPISerializer.extend(
-      JSONAPISerializerMixin,
+      JSONAPIBagaaravelSerializerMixin,
       RelationshipSupportSerializerMixin,
     );
 
@@ -54,7 +54,7 @@ module('Unit | Mixin | relationship-support-serializer', function (hooks) {
     });
 
     let UserSerializer = JSONAPISerializer.extend(
-      JSONAPISerializerMixin,
+      JSONAPIBagaaravelSerializerMixin,
       RelationshipSupportSerializerMixin,
     );
 
@@ -94,7 +94,7 @@ module('Unit | Mixin | relationship-support-serializer', function (hooks) {
     });
 
     let UserSerializer = JSONAPISerializer.extend(
-      JSONAPISerializerMixin,
+      JSONAPIBagaaravelSerializerMixin,
       RelationshipSupportSerializerMixin,
     );
